refactor [#14] : CategoryIntegrationTest 분리
- CategoryIntegrationTest를 extends 해서 CategoryIntegrationTest 에 있는 service들과 공통으로 사용할 메서드들을 만들어서 사용할 수 있도록 할 예정이다.
