+++ backend/src/test/java/com/handong/rebon/integration/category/CategoryReadIntegrationTest.java
@@ -1,0 +1,5 @@
+package com.handong.rebon.integration.category;
+
+public class CategoryReadIntegrationTest extends CategoryIntegrationTest{
+
+}
